Promote log_warnings to the shared plugin test helper
Add the log_warnings loguru-capture fixture to register_plugin_test_fixtures
(imbue.mngr.utils.plugin_testing) so all helper-using plugins get it, and drop
the plugin-local copy from mngr_pi_coding's conftest. Mirrors how
setup_test_mngr_env is exposed both via the helper and (separately) in mngr's
own conftest, so mngr's own log_warnings copy is left untouched for its tests.
