+++ libs/mngr/changelog/mngr-bad-tests-mngr-pi-coding-local.md
@@ -1,0 +1,1 @@
+Added the `log_warnings` loguru-capture fixture to the shared plugin test helper (`register_plugin_test_fixtures` in `imbue.mngr.utils.plugin_testing`), so plugin test suites that register the standard fixtures can assert on emitted warnings without defining their own copy. (Affects test infrastructure only.)
